basic/user-util: getgroups_alloc(): Remove hard-coded initial value of ngroups
Let's call getgroups(0,NULL) also in the first iteration to size the
buffer.  Performance shouldn't be a concern, since the new() and
subsequent getgroups(3) calls are probably mcuh heavier.

This simplifies the code, allows reducing the scope of 'ngroups', and
will permit a simplification of the loop in the following commit.

Signed-off-by: Alejandro Colomar <alx@kernel.org>